Add fix around regex pattern used for matching fixture queries.
The root issue of env variables not working in path was that the
parsePath method was not using the same regex pattern that was used in
parseParam. This divergence caused the env variables to be supported as
params but not in a path.

After adding a test and creating a common util function to extract the
fixture query matches it was found that even the regex pattern that
seemed to work for both cases needed to be tweaked for the case where
multiple fixture queries are used in a row.