Refactor parseQuery to use new fixtureQuery struct.
In parseQuery we were indexing into matched regex groups. This can be
difficult to read and maintain if we ever chose to change the regex
pattern (or go with something more robust entirely).

A new util function has been added which leverages the matchFixtureQuery
util to create an actual struct from the matched regex groups.