Add fix around regex pattern used for matching fixture queries (#434)
* Add fix around regex pattern used for matching fixture queries.

The root issue of env variables not working in path was that the
parsePath method was not using the same regex pattern that was used in
parseParam. This divergence caused the env variables to be supported as
params but not in a path.

After adding a test and creating a common util function to extract the
fixture query matches it was found that even the regex pattern that
seemed to work for both cases needed to be tweaked for the case where
multiple fixture queries are used in a row.

* Refactor parseQuery to use new fixtureQuery struct.

In parseQuery we were indexing into matched regex groups. This can be
difficult to read and maintain if we ever chose to change the regex
pattern (or go with something more robust entirely).

A new util function has been added which leverages the matchFixtureQuery
util to create an actual struct from the matched regex groups.

* Add tests around toFixtureQuery util.

* Update comments