Refactor the way we think about 'serializable' objects with respect to the interactionRecorder and interactionReplayer.
As pointed out by Richard in PR comments, the previous way of storing the serialization logic for a type in a wrapper type is clunky, particularly because
it means that a empty instance needs to be available to perform deserialization. Now, the serialization and deserialization logic are defined in standalone functions (with the
`serialization` and `deserialization` types) and the interactionRecorder and interactionReplayer are intialized with the necessary de/seriaization functions.

This cuts down on the layers of indirection when working with serialization/deserialization code.