Fix bug - switchMode() method on Server now sets cassetteLoaded = false.
Before this, if you called switch mode on a server with a loaded cassette, and then tried to record/replay interactions you could get unexpected errors/crashes in the server. This was because the server wasn't forcing you to load a new cassette after switching mode. Setting cassetteLoaded = false after each mode switch fixes this.