Instead of passing around http.Request and http.Response objects in our handler and playback code, use custom wrapper structs we've defined.
This avoids the messiness of having to work around http.Response's 'stream'-y nature, under which response.Body can only be read once, and if you ever want to re-read the body off of the same http.Response object, you need to do some messy logic to reset the reader.

Now, we immediately convert http.Response and http.Request into our own representation which allows us to read fields as many times as we want, and is a simpler data representation that stills stores all the information we care about.