Support events <= 2017-05-25 without request data
In API versions up to and including 2017-05-25, the event schema included the
request ID that triggered the event as a string in the event object. However,
this ID is only populated for direct Stripe API calls, and is null when the
event was triggered for another reason (e.g., internal Stripe systems, 3DS
verification callbacks, etc.).

This updates `ExtractRequestData` to handle missing request IDs for older API
versions. It also switches from using reflect to a type switch, which requires
less boilerplate.